modify serialization of pipeline
