--- cocos/core/pipeline/common/ubo-stage.ts
@@ -1,138 +1,0 @@
-/**
- * @category pipeline
- */
-
-import { ccclass } from '../../data/class-decorator';
-import { RenderFlow } from '../render-flow';
-import { IRenderStageInfo, RenderStage } from '../render-stage';
-import { RenderView } from '../render-view';
-import { ForwardStagePriority } from '../forward/enum';
-import { RenderContext } from '../render-context';
-import { UBOGlobal} from '../define';
-import { legacyCC } from '../../global-exports';
-import { Mat4, Vec3, Vec4 } from '../../math';
-
-/**
- * @en The ubo stage
- * @zh UBO 处理阶段。
- */
-@ccclass('UBOStage')
-export class UBOStage extends RenderStage {
-
-    public static initInfo: IRenderStageInfo = {
-        name: 'UBOStage',
-        priority: ForwardStagePriority.UBO,
-        renderQueues: [],
-    };
-
-    constructor () {
-        super();
-    }
-
-    public activate (rctx: RenderContext, flow: RenderFlow) {
-        super.activate(rctx, flow);
-    }
-
-    public destroy () {
-    }
-
-    public resize (width: number, height: number) {
-    }
-
-    public rebuild () {
-    }
-
-    public render (rctx: RenderContext, view: RenderView) {
-        const camera = view.camera;
-        const scene = camera.scene!;
-
-        const mainLight = scene.mainLight;
-        const ambient = scene.ambient;
-        const fog = scene.fog;
-        const uboGlobal = rctx.uboGlobal;
-        const fv = uboGlobal.view;
-        const device = rctx.device;
-
-        const shadingWidth = Math.floor(device.width);
-        const shadingHeight = Math.floor(device.height);
-
-        // update UBOGlobal
-        fv[UBOGlobal.TIME_OFFSET] = rctx.cumulativeTime;
-        fv[UBOGlobal.TIME_OFFSET + 1] = rctx.frameTime;
-        fv[UBOGlobal.TIME_OFFSET + 2] = legacyCC.director.getTotalFrames();
-
-        fv[UBOGlobal.SCREEN_SIZE_OFFSET] = device.width;
-        fv[UBOGlobal.SCREEN_SIZE_OFFSET + 1] = device.height;
-        fv[UBOGlobal.SCREEN_SIZE_OFFSET + 2] = 1.0 / fv[UBOGlobal.SCREEN_SIZE_OFFSET];
-        fv[UBOGlobal.SCREEN_SIZE_OFFSET + 3] = 1.0 / fv[UBOGlobal.SCREEN_SIZE_OFFSET + 1];
-
-        fv[UBOGlobal.SCREEN_SCALE_OFFSET] = camera.width / shadingWidth * rctx.shadingScale;
-        fv[UBOGlobal.SCREEN_SCALE_OFFSET + 1] = camera.height / shadingHeight * rctx.shadingScale;
-        fv[UBOGlobal.SCREEN_SCALE_OFFSET + 2] = 1.0 / fv[UBOGlobal.SCREEN_SCALE_OFFSET];
-        fv[UBOGlobal.SCREEN_SCALE_OFFSET + 3] = 1.0 / fv[UBOGlobal.SCREEN_SCALE_OFFSET + 1];
-
-        fv[UBOGlobal.NATIVE_SIZE_OFFSET] = shadingWidth;
-        fv[UBOGlobal.NATIVE_SIZE_OFFSET + 1] = shadingHeight;
-        fv[UBOGlobal.NATIVE_SIZE_OFFSET + 2] = 1.0 / fv[UBOGlobal.NATIVE_SIZE_OFFSET];
-        fv[UBOGlobal.NATIVE_SIZE_OFFSET + 3] = 1.0 / fv[UBOGlobal.NATIVE_SIZE_OFFSET + 1];
-
-        Mat4.toArray(fv, camera.matView, UBOGlobal.MAT_VIEW_OFFSET);
-
-        Mat4.toArray(fv, camera.node.worldMatrix, UBOGlobal.MAT_VIEW_INV_OFFSET);
-        Mat4.toArray(fv, camera.matProj, UBOGlobal.MAT_PROJ_OFFSET);
-        Mat4.toArray(fv, camera.matProjInv, UBOGlobal.MAT_PROJ_INV_OFFSET);
-        Mat4.toArray(fv, camera.matViewProj, UBOGlobal.MAT_VIEW_PROJ_OFFSET);
-        Mat4.toArray(fv, camera.matViewProjInv, UBOGlobal.MAT_VIEW_PROJ_INV_OFFSET);
-        Vec3.toArray(fv, camera.position, UBOGlobal.CAMERA_POS_OFFSET);
-        fv[UBOGlobal.CAMERA_POS_OFFSET + 3] = device.projectionSignY;
-
-        const exposure = camera.exposure;
-        fv[UBOGlobal.EXPOSURE_OFFSET] = exposure;
-        fv[UBOGlobal.EXPOSURE_OFFSET + 1] = 1.0 / exposure;
-        fv[UBOGlobal.EXPOSURE_OFFSET + 2] = rctx.isHDR ? 1.0 : 0.0;
-        fv[UBOGlobal.EXPOSURE_OFFSET + 3] = rctx.fpScale / exposure;
-
-        if (mainLight) {
-            Vec3.toArray(fv, mainLight.direction, UBOGlobal.MAIN_LIT_DIR_OFFSET);
-            Vec3.toArray(fv, mainLight.color, UBOGlobal.MAIN_LIT_COLOR_OFFSET);
-            if (mainLight.useColorTemperature) {
-                const colorTempRGB = mainLight.colorTemperatureRGB;
-                fv[UBOGlobal.MAIN_LIT_COLOR_OFFSET] *= colorTempRGB.x;
-                fv[UBOGlobal.MAIN_LIT_COLOR_OFFSET + 1] *= colorTempRGB.y;
-                fv[UBOGlobal.MAIN_LIT_COLOR_OFFSET + 2] *= colorTempRGB.z;
-            }
-
-            if (rctx.isHDR) {
-                fv[UBOGlobal.MAIN_LIT_COLOR_OFFSET + 3] = mainLight.illuminance * rctx.fpScale;
-            } else {
-                fv[UBOGlobal.MAIN_LIT_COLOR_OFFSET + 3] = mainLight.illuminance * exposure;
-            }
-        } else {
-            Vec3.toArray(fv, Vec3.UNIT_Z, UBOGlobal.MAIN_LIT_DIR_OFFSET);
-            Vec4.toArray(fv, Vec4.ZERO, UBOGlobal.MAIN_LIT_COLOR_OFFSET);
-        }
-
-        const skyColor = ambient.skyColor;
-        if (rctx.isHDR) {
-            skyColor[3] = ambient.skyIllum * rctx.fpScale;
-        } else {
-            skyColor[3] = ambient.skyIllum * exposure;
-        }
-        fv.set(skyColor, UBOGlobal.AMBIENT_SKY_OFFSET);
-
-        fv.set(ambient.groundAlbedo, UBOGlobal.AMBIENT_GROUND_OFFSET);
-
-        fv.set(fog.fogColor, UBOGlobal.GLOBAL_FOG_COLOR_OFFSET);
-
-        fv[UBOGlobal.GLOBAL_FOG_BASE_OFFSET] = fog.fogStart;
-        fv[UBOGlobal.GLOBAL_FOG_BASE_OFFSET + 1] = fog.fogEnd;
-        fv[UBOGlobal.GLOBAL_FOG_BASE_OFFSET + 2] = fog.fogDensity;
-
-        fv[UBOGlobal.GLOBAL_FOG_ADD_OFFSET] = fog.fogTop;
-        fv[UBOGlobal.GLOBAL_FOG_ADD_OFFSET + 1] = fog.fogRange;
-        fv[UBOGlobal.GLOBAL_FOG_ADD_OFFSET + 2] = fog.fogAtten;
-
-        // update ubos
-        rctx.globalBindings.get(UBOGlobal.BLOCK.name)!.buffer!.update(uboGlobal.view);
-    }
-}
