+++ cocos/core/pipeline/render-context.ts
@@ -1,18 +1,21 @@
 import { ccclass, property } from '../data/class-decorator';
 import { Root } from '../root';
 import { GFXDevice, GFXFeature } from '../gfx/device';
-import { RenderPipeline } from './render-pipeline';
-import { IRenderObject, UBOGlobal, IInternalBindingInst, UBOShadow, UNIFORM_ENVIRONMENT, UBOForwardLight} from './define';
-import { GFXBindingType, GFXBufferUsageBit, GFXMemoryUsageBit, GFXFormat,
+import { IRenderObject, UBOGlobal, IInternalBindingInst, UBOShadow,
+    UNIFORM_ENVIRONMENT, UBOForwardLight, RenderPassStage} from './define';
+import { GFXBindingType, GFXBufferUsageBit, GFXMemoryUsageBit, GFXFormat, GFXStoreOp,
     GFXFormatInfos, GFXTextureUsageBit, GFXCommandBufferType, GFXClearFlag } from '../gfx/define';
 import { GFXBuffer } from '../gfx/buffer';
 import { GFXCommandBuffer } from '../gfx/command-buffer';
 import { RenderTextureConfig, MaterialConfig } from './pipeline-serialization';
 import { RenderTexture } from '../assets/render-texture';
-import { Material } from '..';
+import { Material } from '../assets/material';
 import { GFXInputAssembler, IGFXAttribute } from '../gfx/input-assembler';
 import { GFXColorAttachment, GFXDepthStencilAttachment, GFXRenderPass, GFXLoadOp, GFXTextureLayout } from '../gfx';
 import { SKYBOX_FLAG } from '../renderer';
+import { director } from '../director';
+import { RenderView } from './render-view';
+import { Mat4, Vec3, Vec4 } from '../math';
 
 @ccclass('RenderContext')
 export class RenderContext {
@@ -26,9 +29,8 @@ export class RenderContext {
     })
     protected materials: MaterialConfig[] = [];
     protected _materials: Map<string, Material> = new Map<string, Material>();
-    protected root: Root;
-    protected pipeline: RenderPipeline;
-    public device: GFXDevice;
+    protected root: Root | null = null;
+    public device: GFXDevice | null = null;
     public get isHDR () {
         return this._isHDR;
     }
@@ -124,15 +126,15 @@ export class RenderContext {
     }
 
     public get mainWindow () {
-        return this.root.mainWindow;
+        return this.root!.mainWindow;
     }
 
     public get cumulativeTime () {
-        return this.root.cumulativeTime;
+        return this.root!.cumulativeTime;
     }
 
     public get frameTime () {
-        return this.root.frameTime;
+        return this.root!.frameTime;
     }
     /**
      * @en The current frame buffer id for shading
@@ -176,6 +178,14 @@ export class RenderContext {
     public get depthStencilFormat () {
         return this._depthStencilFmt;
     }
+    /**
+     * @en Whether use dynamic batching in this pipeline
+     * @zh 是否启用动态合批。
+     * @readonly
+     */
+    public get useDynamicBatching (): boolean {
+        return this._useDynamicBatching;
+    }
     protected _usePostProcess: boolean = false;
     protected _isHDR: boolean = false;
     protected _useMSAA: boolean = false;
@@ -195,6 +205,8 @@ export class RenderContext {
     protected _quadVB: GFXBuffer | null = null;
     protected _quadIB: GFXBuffer | null = null;
     protected _quadIA: GFXInputAssembler | null = null;
+    protected _useDynamicBatching: boolean = false;
+    protected _renderPasses = new Map<GFXClearFlag, GFXRenderPass>();
     /**
      * @en The list for render objects, only available after the scene culling of the current frame.
      * @zh 渲染对象数组，仅在当前帧的场景剔除完成后有效。
@@ -204,84 +216,17 @@ export class RenderContext {
     public uboGlobal: UBOGlobal = new UBOGlobal();
     public globalBindings: Map<string, IInternalBindingInst> = new Map<string, IInternalBindingInst>();
     public commandBuffers: GFXCommandBuffer[] = [];
-    private _renderPasses = new Map<GFXClearFlag, GFXRenderPass>();
-
-    constructor (root: Root, device: GFXDevice, ppl: RenderPipeline) {
-        this.root = root;
-        this.device = device;
-        this.pipeline = ppl;
-    }
-
-    /**
-     * @en Get the [[RenderTexture]] referred by the name in the current pipeline
-     * @zh 获取当前管线中名字对应的 [[RenderTexture]] 对象
-     * @param name 名字
-     */
-    public getRenderTexture (name: string) {
-        const rt = this._renderTextures.get(name);
-        if (!rt) {
-            console.error('RenderTexture:' + name + ' not found!');
-            return null;
-        }
-        return rt;
-    }
-
-    public getTexture (name: string) {
-        return this.pipeline.getTexture(name);
-    }
-
-    public getFrameBuffer (name: string) {
-        return this.pipeline.getFrameBuffer(name)!;
-    }
-    public getRenderPass (clearFlags: GFXClearFlag): GFXRenderPass {
-        let renderPass = this._renderPasses.get(clearFlags);
-        if (renderPass) { return renderPass; }
-
-        const device = this.device;
-        const colorAttachment = new GFXColorAttachment();
-        const depthStencilAttachment = new GFXDepthStencilAttachment();
-        colorAttachment.format = device.colorFormat;
-        depthStencilAttachment.format = device.depthStencilFormat;
-
-        if (!(clearFlags & GFXClearFlag.COLOR)) {
-            if (clearFlags & SKYBOX_FLAG) {
-                colorAttachment.loadOp = GFXLoadOp.DISCARD;
-            } else {
-                colorAttachment.loadOp = GFXLoadOp.LOAD;
-                colorAttachment.beginLayout = GFXTextureLayout.PRESENT_SRC;
-            }
-        }
-
-        if ((clearFlags & GFXClearFlag.DEPTH_STENCIL) !== GFXClearFlag.DEPTH_STENCIL) {
-            if (!(clearFlags & GFXClearFlag.DEPTH)) depthStencilAttachment.depthLoadOp = GFXLoadOp.LOAD;
-            if (!(clearFlags & GFXClearFlag.STENCIL)) depthStencilAttachment.stencilLoadOp = GFXLoadOp.LOAD;
-            depthStencilAttachment.beginLayout = GFXTextureLayout.DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
-        }
-
-        renderPass = device.createRenderPass({
-            colorAttachments: [colorAttachment],
-            depthStencilAttachment,
-        });
-        this._renderPasses.set(clearFlags, renderPass!);
-
-        return renderPass;
-    }
-
-    public getMaterial (name: string) {
-        const mat = this._materials.get(name);
-        if (!mat) {
-            console.error('Material:' + name + ' not found!');
-            return null;
-        }
-        return mat;
-    }
-
-    public activate () {
-        this.initialize();
-    }
 
     public initialize () {
-        const device = this.device;
+        const device = this.device!;
+
+        if (!this.createQuadInputAssembler()) {
+            return false;
+        }
+
+        if (!this.createUBOs()) {
+            return false;
+        }
 
         // init textures
         for (let i = 0; i < this.renderTextures.length; i++) {
@@ -326,7 +271,7 @@ export class RenderContext {
                     this._isHDR = true;
                 }
             } else if (device.hasFeature(GFXFeature.COLOR_FLOAT) &&
-                       device.hasFeature(GFXFeature.TEXTURE_FLOAT_LINEAR)) {
+                        device.hasFeature(GFXFeature.TEXTURE_FLOAT_LINEAR)) {
                 if (device.hasFeature(GFXFeature.TEXTURE_FLOAT)) {
                     this._colorFmt = GFXFormat.RGBA32F;
                     this._isHDR = true;
@@ -358,6 +303,131 @@ export class RenderContext {
             type: GFXCommandBufferType.PRIMARY,
             queue: device.queue,
         });
+
+        let colorAttachment = new GFXColorAttachment();
+        let depthStencilAttachment = new GFXDepthStencilAttachment();
+        colorAttachment.format = device.colorFormat;
+        depthStencilAttachment.format = device.depthStencilFormat;
+        depthStencilAttachment.depthStoreOp = GFXStoreOp.DISCARD;
+        depthStencilAttachment.stencilStoreOp = GFXStoreOp.DISCARD;
+
+        const windowPass = device.createRenderPass({
+            colorAttachments: [colorAttachment],
+            depthStencilAttachment,
+        });
+        this.addRenderPass(RenderPassStage.DEFAULT, windowPass);
+
+        colorAttachment = new GFXColorAttachment();
+        colorAttachment.format = device.colorFormat;
+        colorAttachment.loadOp = GFXLoadOp.LOAD;
+        colorAttachment.beginLayout = GFXTextureLayout.PRESENT_SRC;
+
+        depthStencilAttachment = new GFXDepthStencilAttachment();
+        depthStencilAttachment.format = device.depthStencilFormat;
+        depthStencilAttachment.depthStoreOp = GFXStoreOp.DISCARD;
+        depthStencilAttachment.stencilStoreOp = GFXStoreOp.DISCARD;
+
+        const uiPass = device.createRenderPass({
+            colorAttachments: [colorAttachment],
+            depthStencilAttachment,
+        });
+        this.addRenderPass(RenderPassStage.UI, uiPass);
+
+    }
+
+    /**
+     * @en Add a render pass.
+     * @zh 添加渲染过程。
+     * @param stage The render stage id
+     * @param renderPass The render pass setting for the stage
+     */
+    public addRenderPass (stage: number, renderPass: GFXRenderPass) {
+        if (renderPass) {
+            this._renderPasses.set(stage, renderPass);
+        }
+    }
+
+    public getRenderTextures () {
+        return this._renderTextures;
+    }
+
+    /**
+     * @en Get the [[RenderTexture]] referred by the name in the current pipeline
+     * @zh 获取当前管线中名字对应的 [[RenderTexture]] 对象
+     * @param name 名字
+     */
+    public getRenderTexture (name: string) {
+        const rt = this._renderTextures.get(name);
+        if (!rt) {
+            console.error('RenderTexture:' + name + ' not found!');
+            return null;
+        }
+        return rt;
+    }
+
+    public addRenderTexture (name: string, rt: RenderTexture) {
+        this._renderTextures.set(name, rt);
+    }
+
+    public getTexture (name: string) {
+        const rt = this.getRenderTexture(name);
+        return rt!.getGFXTexture();
+    }
+
+    public getFrameBuffer (name: string) {
+        const rt = this.getRenderTexture(name);
+        return rt!.window!.framebuffer;
+    }
+
+    public getRenderPass (clearFlags: GFXClearFlag): GFXRenderPass {
+        let renderPass = this._renderPasses.get(clearFlags);
+        if (renderPass) { return renderPass; }
+
+        const device = this.device!;
+        const colorAttachment = new GFXColorAttachment();
+        const depthStencilAttachment = new GFXDepthStencilAttachment();
+        colorAttachment.format = device.colorFormat;
+        depthStencilAttachment.format = device.depthStencilFormat;
+
+        if (!(clearFlags & GFXClearFlag.COLOR)) {
+            if (clearFlags & SKYBOX_FLAG) {
+                colorAttachment.loadOp = GFXLoadOp.DISCARD;
+            } else {
+                colorAttachment.loadOp = GFXLoadOp.LOAD;
+                colorAttachment.beginLayout = GFXTextureLayout.PRESENT_SRC;
+            }
+        }
+
+        if ((clearFlags & GFXClearFlag.DEPTH_STENCIL) !== GFXClearFlag.DEPTH_STENCIL) {
+            if (!(clearFlags & GFXClearFlag.DEPTH)) depthStencilAttachment.depthLoadOp = GFXLoadOp.LOAD;
+            if (!(clearFlags & GFXClearFlag.STENCIL)) depthStencilAttachment.stencilLoadOp = GFXLoadOp.LOAD;
+            depthStencilAttachment.beginLayout = GFXTextureLayout.DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
+        }
+
+        renderPass = device.createRenderPass({
+            colorAttachments: [colorAttachment],
+            depthStencilAttachment,
+        });
+        this._renderPasses.set(clearFlags, renderPass!);
+
+        return renderPass;
+    }
+
+    public getMaterial (name: string) {
+        const mat = this._materials.get(name);
+        if (!mat) {
+            console.error('Material:' + name + ' not found!');
+            return null;
+        }
+        return mat;
+    }
+
+    public activate (root: Root) {
+        this.root = root;
+        this.device = root.device;
+    }
+
+    public sceneCulling (view: RenderView) {
     }
 
     /**
@@ -365,7 +435,7 @@ export class RenderContext {
      * @zh 创建所有 UBO。
      */
     public createUBOs (): boolean {
-        const device = this.device;
+        const device = this.device!;
         if (!this.globalBindings.get(UBOGlobal.BLOCK.name)) {
             const globalUBO = device.createBuffer({
                 usage: GFXBufferUsageBit.UNIFORM | GFXBufferUsageBit.TRANSFER_DST,
@@ -404,12 +474,110 @@ export class RenderContext {
         return true;
     }
 
+    public updateUBO (view: RenderView) {
+        const camera = view.camera;
+        const scene = camera.scene!;
+
+        const mainLight = scene.mainLight;
+        const ambient = scene.ambient;
+        const fog = scene.fog;
+        const uboGlobal = this.uboGlobal;
+        const fv = uboGlobal.view;
+        const device = this.device!;
+
+        const shadingWidth = Math.floor(device.width);
+        const shadingHeight = Math.floor(device.height);
+
+        // update UBOGlobal
+        fv[UBOGlobal.TIME_OFFSET] = this.cumulativeTime;
+        fv[UBOGlobal.TIME_OFFSET + 1] = this.frameTime;
+        fv[UBOGlobal.TIME_OFFSET + 2] = director.getTotalFrames();
+
+        fv[UBOGlobal.SCREEN_SIZE_OFFSET] = device.width;
+        fv[UBOGlobal.SCREEN_SIZE_OFFSET + 1] = device.height;
+        fv[UBOGlobal.SCREEN_SIZE_OFFSET + 2] = 1.0 / fv[UBOGlobal.SCREEN_SIZE_OFFSET];
+        fv[UBOGlobal.SCREEN_SIZE_OFFSET + 3] = 1.0 / fv[UBOGlobal.SCREEN_SIZE_OFFSET + 1];
+
+        fv[UBOGlobal.SCREEN_SCALE_OFFSET] = camera.width / shadingWidth * this.shadingScale;
+        fv[UBOGlobal.SCREEN_SCALE_OFFSET + 1] = camera.height / shadingHeight * this.shadingScale;
+        fv[UBOGlobal.SCREEN_SCALE_OFFSET + 2] = 1.0 / fv[UBOGlobal.SCREEN_SCALE_OFFSET];
+        fv[UBOGlobal.SCREEN_SCALE_OFFSET + 3] = 1.0 / fv[UBOGlobal.SCREEN_SCALE_OFFSET + 1];
+
+        fv[UBOGlobal.NATIVE_SIZE_OFFSET] = shadingWidth;
+        fv[UBOGlobal.NATIVE_SIZE_OFFSET + 1] = shadingHeight;
+        fv[UBOGlobal.NATIVE_SIZE_OFFSET + 2] = 1.0 / fv[UBOGlobal.NATIVE_SIZE_OFFSET];
+        fv[UBOGlobal.NATIVE_SIZE_OFFSET + 3] = 1.0 / fv[UBOGlobal.NATIVE_SIZE_OFFSET + 1];
+
+        Mat4.toArray(fv, camera.matView, UBOGlobal.MAT_VIEW_OFFSET);
+
+        Mat4.toArray(fv, camera.node.worldMatrix, UBOGlobal.MAT_VIEW_INV_OFFSET);
+        Mat4.toArray(fv, camera.matProj, UBOGlobal.MAT_PROJ_OFFSET);
+        Mat4.toArray(fv, camera.matProjInv, UBOGlobal.MAT_PROJ_INV_OFFSET);
+        Mat4.toArray(fv, camera.matViewProj, UBOGlobal.MAT_VIEW_PROJ_OFFSET);
+        Mat4.toArray(fv, camera.matViewProjInv, UBOGlobal.MAT_VIEW_PROJ_INV_OFFSET);
+        Vec3.toArray(fv, camera.position, UBOGlobal.CAMERA_POS_OFFSET);
+        let projectionSignY = device.screenSpaceSignY;
+        if (view.window.hasOffScreenAttachments) {
+            projectionSignY *= device.UVSpaceSignY; // need flipping if drawing on render targets
+        }
+        fv[UBOGlobal.CAMERA_POS_OFFSET + 3] = projectionSignY;
+
+        const exposure = camera.exposure;
+        fv[UBOGlobal.EXPOSURE_OFFSET] = exposure;
+        fv[UBOGlobal.EXPOSURE_OFFSET + 1] = 1.0 / exposure;
+        fv[UBOGlobal.EXPOSURE_OFFSET + 2] = this.isHDR ? 1.0 : 0.0;
+        fv[UBOGlobal.EXPOSURE_OFFSET + 3] = this.fpScale / exposure;
+
+        if (mainLight) {
+            Vec3.toArray(fv, mainLight.direction, UBOGlobal.MAIN_LIT_DIR_OFFSET);
+            Vec3.toArray(fv, mainLight.color, UBOGlobal.MAIN_LIT_COLOR_OFFSET);
+            if (mainLight.useColorTemperature) {
+                const colorTempRGB = mainLight.colorTemperatureRGB;
+                fv[UBOGlobal.MAIN_LIT_COLOR_OFFSET] *= colorTempRGB.x;
+                fv[UBOGlobal.MAIN_LIT_COLOR_OFFSET + 1] *= colorTempRGB.y;
+                fv[UBOGlobal.MAIN_LIT_COLOR_OFFSET + 2] *= colorTempRGB.z;
+            }
+
+            if (this.isHDR) {
+                fv[UBOGlobal.MAIN_LIT_COLOR_OFFSET + 3] = mainLight.illuminance * this.fpScale;
+            } else {
+                fv[UBOGlobal.MAIN_LIT_COLOR_OFFSET + 3] = mainLight.illuminance * exposure;
+            }
+        } else {
+            Vec3.toArray(fv, Vec3.UNIT_Z, UBOGlobal.MAIN_LIT_DIR_OFFSET);
+            Vec4.toArray(fv, Vec4.ZERO, UBOGlobal.MAIN_LIT_COLOR_OFFSET);
+        }
+
+        const skyColor = ambient.skyColor;
+        if (this.isHDR) {
+            skyColor[3] = ambient.skyIllum * this.fpScale;
+        } else {
+            skyColor[3] = ambient.skyIllum * exposure;
+        }
+        fv.set(skyColor, UBOGlobal.AMBIENT_SKY_OFFSET);
+
+        fv.set(ambient.groundAlbedo, UBOGlobal.AMBIENT_GROUND_OFFSET);
+
+        fv.set(fog.fogColor, UBOGlobal.GLOBAL_FOG_COLOR_OFFSET);
+
+        fv[UBOGlobal.GLOBAL_FOG_BASE_OFFSET] = fog.fogStart;
+        fv[UBOGlobal.GLOBAL_FOG_BASE_OFFSET + 1] = fog.fogEnd;
+        fv[UBOGlobal.GLOBAL_FOG_BASE_OFFSET + 2] = fog.fogDensity;
+
+        fv[UBOGlobal.GLOBAL_FOG_ADD_OFFSET] = fog.fogTop;
+        fv[UBOGlobal.GLOBAL_FOG_ADD_OFFSET + 1] = fog.fogRange;
+        fv[UBOGlobal.GLOBAL_FOG_ADD_OFFSET + 2] = fog.fogAtten;
+
+        // update ubos
+        this.globalBindings.get(UBOGlobal.BLOCK.name)!.buffer!.update(uboGlobal.view);
+    }
+
     /**
      * @en Create input assembler for quad
      * @zh 创建四边形输入汇集器。
      */
     public createQuadInputAssembler (): boolean {
-        const device = this.device;
+        const device = this.device!;
         // create vertex buffer
 
         const vbStride = Float32Array.BYTES_PER_ELEMENT * 4;
@@ -508,16 +676,6 @@ export class RenderContext {
     }
 
     /**
-     * @en Swap the frame buffer.
-     * @zh 交换帧缓冲。
-     */
-    public swapFBOs () {
-        const temp = this._curIdx;
-        this._curIdx = this._prevIdx;
-        this._prevIdx = temp;
-    }
-
-    /**
      * @en Destroy all UBOs
      * @zh 销毁全部 UBO。
      */
@@ -533,8 +691,23 @@ export class RenderContext {
             this.globalBindings.delete(UBOShadow.BLOCK.name);
         }
     }
+
     public destroy () {
         this.destroyUBOs();
         this.destroyQuadInputAssembler();
+
+        const rtIter = this._renderTextures.values();
+        let rtRes = rtIter.next();
+        while (!rtRes.done) {
+            rtRes.value.destroy();
+            rtRes = rtIter.next();
+        }
+
+        const rpIter = this._renderPasses.values();
+        let rpRes = rpIter.next();
+        while (!rpRes.done) {
+            rpRes.value.destroy();
+            rpRes = rpIter.next();
+        }
     }
 };
